+++ src/storage/storage_transformer/usage_log.rs
@@ -5,14 +5,19 @@ use std::{
     sync::{Arc, Mutex},
 };
 
+use async_trait::async_trait;
+
 use crate::{
     array::MaybeBytes,
     byte_range::ByteRange,
     metadata::Metadata,
     storage::{
-        ListableStorage, ListableStorageTraits, ReadableListableStorage, ReadableStorage,
-        ReadableStorageTraits, StorageError, StoreKey, StoreKeyRange, StoreKeyStartValue,
-        StoreKeys, StoreKeysPrefixes, StorePrefix, WritableStorage, WritableStorageTraits,
+        AsyncListableStorage, AsyncListableStorageTraits, AsyncReadableListableStorage,
+        AsyncReadableStorage, AsyncReadableStorageTraits, AsyncWritableStorage,
+        AsyncWritableStorageTraits, ListableStorage, ListableStorageTraits,
+        ReadableListableStorage, ReadableStorage, ReadableStorageTraits, StorageError, StoreKey,
+        StoreKeyRange, StoreKeyStartValue, StoreKeys, StoreKeysPrefixes, StorePrefix,
+        WritableStorage, WritableStorageTraits,
     },
 };
 
@@ -72,6 +77,38 @@ impl StorageTransformerExtension for UsageLogStorageTransformer {
         &self,
         storage: ReadableListableStorage<'a>,
     ) -> ReadableListableStorage<'a> {
+        self.create_transformer(storage)
+    }
+
+    /// Create an asynchronous readable transformer.
+    fn create_async_readable_transformer<'a>(
+        &'a self,
+        storage: AsyncReadableStorage<'a>,
+    ) -> AsyncReadableStorage<'a> {
+        self.create_transformer(storage)
+    }
+
+    /// Create an asynchronous writable transformer.
+    fn create_async_writable_transformer<'a>(
+        &'a self,
+        storage: AsyncWritableStorage<'a>,
+    ) -> AsyncWritableStorage<'a> {
+        self.create_transformer(storage)
+    }
+
+    /// Create an asynchronous listable transformer.
+    fn create_async_listable_transformer<'a>(
+        &'a self,
+        storage: AsyncListableStorage<'a>,
+    ) -> AsyncListableStorage<'a> {
+        self.create_transformer(storage)
+    }
+
+    /// Create an asynchronous readable and listable transformer.
+    fn create_async_readable_listable_transformer<'a>(
+        &'a self,
+        storage: AsyncReadableListableStorage<'a>,
+    ) -> AsyncReadableListableStorage<'a> {
         self.create_transformer(storage)
     }
 }
@@ -235,5 +272,164 @@ impl<TStorage: ?Sized + WritableStorageTraits> WritableStorageTraits
             (self.prefix_func)()
         )?;
         self.storage.erase_prefix(prefix)
+    }
+}
+
+#[async_trait]
+impl<TStorage: ?Sized + AsyncReadableStorageTraits> AsyncReadableStorageTraits
+    for UsageLogStorageTransformerImpl<TStorage>
+{
+    async fn get(&self, key: &StoreKey) -> Result<MaybeBytes, StorageError> {
+        let result = self.storage.get(key).await;
+        writeln!(
+            self.handle.lock().unwrap(),
+            "{}get({key:?}) -> len={:?}",
+            (self.prefix_func)(),
+            result
+                .as_ref()
+                .map(|v| v.as_ref().map_or(0, std::vec::Vec::len))
+        )?;
+        result
+    }
+
+    async fn get_partial_values_key(
+        &self,
+        key: &StoreKey,
+        byte_ranges: &[ByteRange],
+    ) -> Result<Option<Vec<Vec<u8>>>, StorageError> {
+        let _ = writeln!(
+            self.handle.lock().unwrap(),
+            "{}get_partial_values_key({key}, {byte_ranges:?})",
+            (self.prefix_func)()
+        );
+        self.storage.get_partial_values_key(key, byte_ranges).await
+    }
+
+    async fn get_partial_values(
+        &self,
+        key_ranges: &[StoreKeyRange],
+    ) -> Result<Vec<MaybeBytes>, StorageError> {
+        let _ = writeln!(
+            self.handle.lock().unwrap(),
+            "{}get_partial_values({key_ranges:?})",
+            (self.prefix_func)()
+        );
+        self.storage.get_partial_values(key_ranges).await
+    }
+
+    async fn size(&self) -> Result<u64, StorageError> {
+        let size = self.storage.size().await;
+        let _ = writeln!(
+            self.handle.lock().unwrap(),
+            "{}size() -> {size:?}",
+            (self.prefix_func)()
+        );
+        size
+    }
+
+    async fn size_prefix(&self, prefix: &StorePrefix) -> Result<u64, StorageError> {
+        let size: Result<u64, StorageError> = self.storage.size_prefix(prefix).await;
+        let _ = writeln!(
+            self.handle.lock().unwrap(),
+            "{}size_prefix({prefix}) -> {size:?}",
+            (self.prefix_func)()
+        );
+        size
+    }
+
+    async fn size_key(&self, key: &StoreKey) -> Result<Option<u64>, StorageError> {
+        let size = self.storage.size_key(key).await;
+        let _ = writeln!(
+            self.handle.lock().unwrap(),
+            "{}size_key({key}) -> {size:?}",
+            (self.prefix_func)()
+        );
+        size
+    }
+}
+
+#[async_trait]
+impl<TStorage: ?Sized + AsyncListableStorageTraits> AsyncListableStorageTraits
+    for UsageLogStorageTransformerImpl<TStorage>
+{
+    async fn list(&self) -> Result<StoreKeys, StorageError> {
+        writeln!(
+            self.handle.lock().unwrap(),
+            "{}list()",
+            (self.prefix_func)()
+        )?;
+        self.storage.list().await
+    }
+
+    async fn list_prefix(&self, prefix: &StorePrefix) -> Result<StoreKeys, StorageError> {
+        writeln!(
+            self.handle.lock().unwrap(),
+            "{}list_prefix({prefix:?})",
+            (self.prefix_func)()
+        )?;
+        self.storage.list_prefix(prefix).await
+    }
+
+    async fn list_dir(&self, prefix: &StorePrefix) -> Result<StoreKeysPrefixes, StorageError> {
+        writeln!(
+            self.handle.lock().unwrap(),
+            "{}list_dir({prefix:?})",
+            (self.prefix_func)()
+        )?;
+        self.storage.list_dir(prefix).await
+    }
+}
+
+#[async_trait]
+impl<TStorage: ?Sized + AsyncWritableStorageTraits> AsyncWritableStorageTraits
+    for UsageLogStorageTransformerImpl<TStorage>
+{
+    async fn set(&self, key: &StoreKey, value: &[u8]) -> Result<(), StorageError> {
+        writeln!(
+            self.handle.lock().unwrap(),
+            "{}set({key:?}, len={})",
+            (self.prefix_func)(),
+            value.len()
+        )?;
+        self.storage.set(key, value).await
+    }
+
+    async fn set_partial_values(
+        &self,
+        key_start_values: &[StoreKeyStartValue],
+    ) -> Result<(), StorageError> {
+        writeln!(
+            self.handle.lock().unwrap(),
+            "{}set_partial_values({key_start_values:?}",
+            (self.prefix_func)()
+        )?;
+        self.storage.set_partial_values(key_start_values).await
+    }
+
+    async fn erase(&self, key: &StoreKey) -> Result<bool, StorageError> {
+        writeln!(
+            self.handle.lock().unwrap(),
+            "{}erase({key:?}",
+            (self.prefix_func)()
+        )?;
+        self.storage.erase(key).await
+    }
+
+    async fn erase_values(&self, keys: &[StoreKey]) -> Result<bool, StorageError> {
+        writeln!(
+            self.handle.lock().unwrap(),
+            "{}erase_values({keys:?}",
+            (self.prefix_func)()
+        )?;
+        self.storage.erase_values(keys).await
+    }
+
+    async fn erase_prefix(&self, prefix: &StorePrefix) -> Result<bool, StorageError> {
+        writeln!(
+            self.handle.lock().unwrap(),
+            "{}erase_prefix({prefix:?}",
+            (self.prefix_func)()
+        )?;
+        self.storage.erase_prefix(prefix).await
     }
 }